requirements.txt
begins
flake8
ipython
pydeepl
nltk
begins
